Merge NC target data into `target-data/` files
Utilized the `NC_clean` function to generate ground truth data set and
manipulated it into the formats required for the files in the
`target-data/` directory. Added a secondary function,
`merge_and_write_csv`, to merge the updated data into the existing file
and then rewrite it. Currently does this by just removing data
corresponding to locations present in the new data, could be updated
more intelligently with a merge of some kind if the whole new data set
for a location is not present.