Add `sync` cron job to `nc-data.yaml`
Added a `sync` job to the `nc-data.yaml` workflow that runs on a cron
schedule of '30 4 * * 3'. This corresponds to 2330 Tuesday EST before
spring forward (2026-03-08) and 0030 Wednesday EST after spring forward.
The job simply syncs the `ACCIDDA/flu-metrocast-nc` fork with the
`reichlab/flu-metrocast` upstream to clean up the pull request for a
data update on Wednesdays. The jobs now filter both on the repository
and on the event name to keep them contained to the same file.